% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_na.R
\name{make_na}
\alias{make_na}
\title{Recode missing indicators to \code{NA}}
\usage{
make_na(x, misvals)
}
\arguments{
\item{x}{A vector}

\item{misvals}{The values to be replaced with \code{NA}}
}
\value{
Returns \code{x} with any indicies originally containing a value from \code{misvals} replaced with \code{NA}
}
\description{
Replaces supplied numeric values with \code{NA}
}
\examples{
table(nhis2018sm$CHLYR, useNA='always')
nhis2018$cholyr_new = make_na(x=nhis2018sm$CHLYR, misvals=c(7,8,9))
table(nhis2018$cholyr_new)

}
