% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhis2018.R
\docType{data}
\name{nhis2018}
\alias{nhis2018}
\title{2018 National Health Interview Survey (NHIS)}
\format{A data frame with 25417 rows and 742 columns}
\usage{
nhis2018
}
\description{
From the NHIS website: The National Health Interview Survey (NHIS) is the principal source of information on the health of the civilian noninstitutionalized population of the United States and is one of the major data collection programs of the National Center for Health Statistics (NCHS) which is part of the Centers for Disease Control and Prevention (CDC).
}
\details{
This data set contains the Sample Adult data from the 2018 NHIS.

For more information, see the NHIS 2018 webapage: \url{https://www.cdc.gov/nchs/nhis/nhis_2018_data_release.htm}. In particular, the variable layout and variable frequency files will be important.
}
\seealso{
\code{\link{nhis2018sm}}
}
\keyword{datasets}
