% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhis2018sm.R
\docType{data}
\name{nhis2018sm}
\alias{nhis2018sm}
\title{Subset of the 2018 National Health Interview Survey (NHIS)}
\format{A data frame with 500 rows and 742 columns}
\usage{
nhis2018sm
}
\description{
A random sample of 500 rows of the 2018 NHIS survey Sample Adult data set.
}
\details{
For more information, see the NHIS 2018 webapage: \url{https://www.cdc.gov/nchs/nhis/nhis_2018_data_release.htm}. In particular, the variable layout and variable frequency files will be important.
}
\seealso{
\code{\link{nhis2018}}
}
\keyword{datasets}
