% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsduh2018.R
\docType{data}
\name{nsduh2018}
\alias{nsduh2018}
\title{2018 National Survey on Drug Use and Health (NSDUH)}
\format{A data frame with 56313 rows and 2691 columns}
\usage{
nsduh2018
}
\description{
From the NSDUH website: The National Survey on Drug Use and Health (NSDUH) series, formerly titled National Household Survey on Drug Abuse, is a major source of statistical information on the use of illicit drugs, alcohol, and tobacco and on mental health issues among members of the U.S. civilian, non-institutional population aged 12 or older. The survey tracks trends in specific substance use and mental illness measures and assesses the consequences of these conditions by examining mental and/or substance use disorders and treatment for these disorders.
}
\details{
For more information, see the NSDUH 2018 webapage: \url{https://www.datafiles.samhsa.gov/study-dataset/national-survey-drug-use-and-health-2018-nsduh-2018-ds0001-nid18758}. In particular, the survey codebook will be important.
}
\seealso{
\code{\link{nsduh2018sm}}
}
\keyword{datasets}
