% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsduh2018sm.R
\docType{data}
\name{nsduh2018sm}
\alias{nsduh2018sm}
\title{Subset of the 2018 National Survey on Drug Use and Health (NSDUH)}
\format{A data frame with 1000 rows and 2691 columns}
\usage{
nsduh2018sm
}
\description{
A random sample of 1000 rows from the 2018 NSDUH survey
}
\details{
For more information, see the NSDUH 2018 webapage: \url{https://www.datafiles.samhsa.gov/study-dataset/national-survey-drug-use-and-health-2018-nsduh-2018-ds0001-nid18758}. In particular, the survey codebook will be important.
}
\seealso{
\code{\link{nsduh2018}}
}
\keyword{datasets}
